+++ clientes/src/main/java/com/sairapuentes/clientes/service/IClienteServicio.java
@@ -7,8 +7,8 @@ import java.util.List;
 
 public interface IClienteServicio {
     List<ClienteResponse> findAll();
-    ClienteResponse findById(int id);
+    ClienteResponse findById(long id);
     ClienteResponse save(ClienteRequest request);
-    ClienteResponse update(int id, ClienteRequest request);
-    void eliminar(int id);
+    ClienteResponse update(long id, ClienteRequest request);
+    void eliminar(long id);
 }
